Switch to internal ftp server from test.untangle.com

+++ hier/usr/lib/python2.7/tests/virus_blocker_tests.py
@@ -3,6 +3,7 @@ import time
 import sys
 import os
 import platform
+import subprocess
 from jsonrpc import ServiceProxy
 from jsonrpc import JSONRPCException
 from global_functions import uvmContext
@@ -98,8 +99,11 @@ class VirusBlockTests(VirusBlockerBaseTests):
 
     # test the cloud scanner with ftp using our special small test virus
     def test_250_ftpCloudSmallBlocked(self):
+        ftp_result = subprocess.call(["ping","-c","1",global_functions.ftpServer ],stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+        if (ftp_result != 0):
+            raise unittest2.SkipTest("FTP server not available")
         remote_control.runCommand("rm -f /tmp/temp_250_ftpVirusBlocked_file")
-        result = remote_control.runCommand("wget -q -O /tmp/temp_250_ftpVirusBlocked_file ftp://test.untangle.com/test/UntangleVirus.exe")
+        result = remote_control.runCommand("wget -q -O /tmp/temp_250_ftpVirusBlocked_file ftp://" + global_functions.ftpServer + "/test/UntangleVirus.exe")
         assert (result == 0)
         md5TestNum = remote_control.runCommand("\"md5sum /tmp/temp_250_ftpVirusBlocked_file | awk '{print $1}'\"", stdout=True)
         print "md5SmallFile <%s> vs md5TestNum <%s>" % (md5SmallFile, md5TestNum)
@@ -107,8 +111,11 @@ class VirusBlockTests(VirusBlockerBaseTests):
 
     # test the cloud scanner with ftp using our special large test virus
     def test_260_ftpCloudLargeBlocked(self):
+        ftp_result = subprocess.call(["ping","-c","1",global_functions.ftpServer ],stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+        if (ftp_result != 0):
+            raise unittest2.SkipTest("FTP server not available")
         remote_control.runCommand("rm -f /tmp/temp_260_ftpVirusBlocked_file")
-        result = remote_control.runCommand("wget -q -O /tmp/temp_260_ftpVirusBlocked_file ftp://test.untangle.com/test/UntangleLargeVirus.exe")
+        result = remote_control.runCommand("wget -q -O /tmp/temp_260_ftpVirusBlocked_file ftp://" + global_functions.ftpServer + "/test/UntangleLargeVirus.exe")
         assert (result == 0)
         md5TestNum = remote_control.runCommand("\"md5sum /tmp/temp_260_ftpVirusBlocked_file | awk '{print $1}'\"", stdout=True)
         print "md5LargeFile <%s> vs md5TestNum <%s>" % (md5LargeFile, md5TestNum)
